refactor: Split vault-tools.ts into one file per tool (#510)
Move the 1,227-line monolith into src/tools/vault/ with one file per
tool class, a shared utils module for path resolution helpers, and a
barrel index that preserves all existing exports.
